pyproject.toml
[project]
name = "project-blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "google-auth>=2.40.3",
    "gspread>=6.2.1",
    "httpx>=0.28.1",
    "mcp[cli]>=1.12.1",
    "oauth2client>=4.1.3",
]
